Moved all opencl work into a sub module
Signed-off by: David Rowenhorst <david.rowenhorst@nrl.navy.mil>
